feat: add field prop to Type column when define a struct type in sheet
